refactor(2161): Move create and update user action widgets to specific folders
