+++ lib/pages/user_action/user_action_detail_page.dart
@@ -6,8 +6,8 @@ import 'package:pass_emploi_app/features/user_action/commentaire/list/action_com
 import 'package:pass_emploi_app/features/user_action/delete/user_action_delete_actions.dart';
 import 'package:pass_emploi_app/features/user_action/update/user_action_update_actions.dart';
 import 'package:pass_emploi_app/models/user_action.dart';
-import 'package:pass_emploi_app/pages/update_user_action_form/update_user_action_form.dart';
 import 'package:pass_emploi_app/pages/user_action/action_commentaires_page.dart';
+import 'package:pass_emploi_app/pages/user_action/update/update_user_action_form.dart';
 import 'package:pass_emploi_app/presentation/display_state.dart';
 import 'package:pass_emploi_app/presentation/user_action/commentaires/action_commentaire_view_model.dart';
 import 'package:pass_emploi_app/presentation/user_action/user_action_details_view_model.dart';
